Make Metadata() return value instead of pointer
A pointer to the Metadata struct used by the reader could cause the
reader to fail with strange errors if the values in the metadata
were changed. Although Metadata is a bit big for returning by value,
I have a hard time imaging a use case where there performance difference
would actually matter.